+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.3] - 2025-11-19
+
+### Changed
+
+  * Allow uppercase letters in datastream ids
+   
 ## [0.7.2] - 2025-11-11
 
   * Add sip.utils.is_bag function